+++ src/shape.h
@@ -23,35 +23,35 @@ public:
 // tests intersections
 class Shape {
 public:
-	virtual bool intersect(const Ray& ray, float& thit, IntersectionInfo& local) = 0;
+	virtual bool intersect(const Ray& ray, double& thit, IntersectionInfo& local) = 0;
 	virtual bool intersectP(Ray& ray) = 0; // TODO what does this do???
-	Shape(Vec d, Vec s, Vec e, Vec a, float shiny);
+	Shape(Vec d, Vec s, Vec e, Vec a, double shiny);
 	virtual ~Shape() = default;
 
 	// TODO
 	Vec m_diffuse, m_specular, m_emission, ambient;
-	float m_shininess;
+	double m_shininess;
 };
 
 class Triangle : public Shape {
 public:
-	bool intersect(const Ray& ray, float& thit, IntersectionInfo& local) override;
+	bool intersect(const Ray& ray, double& thit, IntersectionInfo& local) override;
 	bool intersectP(Ray& ray) override;
-	Triangle(Vec p1, Vec p2, Vec p3, Mat4 t, Vec d, Vec s, Vec e, Vec a, float shiny);
+	Triangle(Vec p1, Vec p2, Vec p3, Mat4 t, Vec d, Vec s, Vec e, Vec a, double shiny);
 
 private:
-	void barycentric(Vec p, float &u, float &v, float &w);
+	void barycentric(Vec p, double &u, double &v, double &w);
 	Vec m_v1, m_v2, m_v3; // vertices
 };
 
 class Sphere : public Shape {
 public:
-	bool intersect(const Ray& ray, float& thit, IntersectionInfo& local) override;
+	bool intersect(const Ray& ray, double& thit, IntersectionInfo& local) override;
 	bool intersectP(Ray& ray) override;
-	Sphere(float cx, float cy, float cz, float r, Mat4 t, Vec d, Vec s, Vec e, Vec a, float shiny);
+	Sphere(double cx, double cy, double cz, double r, Mat4 t, Vec d, Vec s, Vec e, Vec a, double shiny);
 private:
 	Vec m_center;
-	float m_radius;
+	double m_radius;
 	Mat4 m_transformation, m_inverse, m_inverse_transpose;
 };
